Refactoring(server-info): specified exception type for str2int
