+++ netutils_linux_hardware/cpu.py
@@ -2,7 +2,7 @@
 
 from netutils_linux_hardware.grade import Grade
 from netutils_linux_hardware.parser import Parser, YAMLLike
-from netutils_linux_hardware.rater_math import extract
+from netutils_linux_hardware.rate_math import extract
 from netutils_linux_hardware.subsystem import Subsystem
