+++ netutils_linux_hardware/disk.py
@@ -6,7 +6,7 @@ import yaml
 
 from netutils_linux_hardware.grade import Grade
 from netutils_linux_hardware.parser import Parser
-from netutils_linux_hardware.rater_math import extract
+from netutils_linux_hardware.rate_math import extract
 from netutils_linux_hardware.subsystem import Subsystem
