+++ netutils_linux_hardware/net.py
@@ -6,7 +6,7 @@ from six import iteritems, print_
 from netutils_linux_hardware.grade import Grade
 from netutils_linux_hardware.interrupts import IRQQueueCounter
 from netutils_linux_hardware.parser import Parser, YAMLLike
-from netutils_linux_hardware.rater_math import extract
+from netutils_linux_hardware.rate_math import extract
 from netutils_linux_hardware.subsystem import Subsystem
